added interfaces to service
