+++ src/main/java/com/best_store/right_bite/service/catalog/BrandServiceImpl.java
@@ -1,0 +1,29 @@
+package com.best_store.right_bite.service.catalog;
+
+import com.best_store.right_bite.dto.catalog.BrandDTO;
+import com.best_store.right_bite.mapper.catalog.BaseMapper;
+import com.best_store.right_bite.model.catalog.Brand;
+import com.best_store.right_bite.repository.catalog.BrandRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+public class BrandServiceImpl implements BrandService {
+
+    private final BrandRepository brandRepository;
+
+    @Qualifier("BrandEntityToDTO")
+    private final BaseMapper<Brand, BrandDTO> brandEntityToDTOMapper;
+
+    public List<BrandDTO> getAllBrands() {
+        return  brandRepository.findAll().stream().map(brandEntityToDTOMapper::map).collect(Collectors.toList());
+    }
+
+
+
+}
